Fix compilation warnings.

+++ Source/GenerationsRaytracing.X64/DescriptorHeap.cpp
@@ -53,10 +53,10 @@ uint32_t DescriptorHeap::getIncrementSize() const
 
 uint32_t DescriptorHeap::getIndex(D3D12_CPU_DESCRIPTOR_HANDLE cpuHandle) const
 {
-    return (cpuHandle.ptr - m_cpuDescriptorHandle.ptr) / m_incrementSize;
+    return static_cast<uint32_t>(cpuHandle.ptr - m_cpuDescriptorHandle.ptr) / m_incrementSize;
 }
 
 uint32_t DescriptorHeap::getIndex(D3D12_GPU_DESCRIPTOR_HANDLE gpuHandle) const
 {
-    return (gpuHandle.ptr - m_gpuDescriptorHandle.ptr) / m_incrementSize;
+    return static_cast<uint32_t>(gpuHandle.ptr - m_gpuDescriptorHandle.ptr) / m_incrementSize;
 }
